django/requirements.txt
# Django and GeoDjango
Django>=5.0,<6.0

# PostGIS support
psycopg2-binary>=2.9.9

# GDAL/GEOS libraries for GeoDjango (system-level dependencies required)
# On Docker: These should be installed in the base image
# On local: Use system package manager or conda

# Additional utilities (optional)
python-decouple>=3.8

